Implemented basic movement! Added missing initialisation command Strip grammar from incoming packets before attempting to parse the contents! Parse out numbers correctly Updated UWP app Reactoring

+++ OpenSynScanWifi/OpenSynScanWifi/Commands/IMountInitialisationCommandParser.cs
@@ -1,16 +1,15 @@
 ﻿using OpenSynScanWifi.Annotations;
-using OpenSynScanWifi.Models;
 
 namespace OpenSynScanWifi.Commands
 {
 	public interface IMountInitialisationCommandParser
 	{
-		long ParseCountsPerRevolutionRepsonse([NotNull] byte[] response, long mcVersion);
+		double ParseCountsPerRevolutionRepsonse([NotNull] byte[] response, long mcVersion);
 
-		long ParseHighSpeedRatioResponse([NotNull] byte[] response);
+		double ParseHighSpeedRatioResponse([NotNull] byte[] response);
 
 		long ParseMotorBoardResponse([CanBeNull] byte[] response);
 
-		long ParseTimerInterruptFreqResponse([NotNull] byte[] response);
+		double ParseTimerInterruptFreqResponse([NotNull] byte[] response);
 	}
 }
